all the images were moved into src/assets/images
