+++ src/assets/icons/CleanInput.tsx
