+++ src/assets/icons/WatchPassCrossedOut.tsx
