+++ src/assets/icons/checkEmailIcon.tsx
