+++ src/components/ui/header/header.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 
 import s from './header.module.scss'
 
-import { Logo } from '@/components/assets/icons/Logo.tsx'
+import { Logo } from '@/assets/icons/Logo.tsx'
 import { Button, Typography } from '@/components/ui'
 import { Avatar } from '@/components/ui/avatar'
